fix(#1): Cast head coordinates to int for grid alignment
- Fixes Issue #1: head was initialized with float (self.w/2, self.h/2)
- Cast to int to ensure grid-aligned coordinates
- Prevents collision check failures and misalignment with food

+++ game.py
@@ -42,7 +42,7 @@ class SnakeGameAI:
         # init game state
         self.direction = Direction.RIGHT
 
-        self.head = Point(self.w/2, self.h/2)
+        self.head = Point(int(self.w/2), int(self.h/2))
         self.snake = [self.head,
                       Point(self.head.x-BLOCK_SIZE, self.head.y),
                       Point(self.head.x-(2*BLOCK_SIZE), self.head.y)]
